+++ integration-test/src/test/java/org/apache/iotdb/db/it/aligned/IoTDBInsertAlignedValues2IT.java
@@ -35,6 +35,8 @@ import java.sql.SQLException;
 import java.sql.Statement;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 @RunWith(IoTDBTestRunner.class)
 @Category({LocalStandaloneIT.class, ClusterIT.class})
@@ -123,6 +125,46 @@ public class IoTDBInsertAlignedValues2IT {
       statement.execute("insert into root.sg.d1(time, s1,s2) aligned values(1,'aa','bb')");
       statement.execute("flush");
       statement.execute("insert into root.sg.d1(time, s1,s2) aligned values(1,'aa','bb')");
+    }
+  }
+
+  @Test
+  public void testInsertComplexAlignedValues() throws SQLException {
+    try (Connection connection = EnvFactory.getEnv().getConnection();
+        Statement statement = connection.createStatement()) {
+      statement.addBatch("create aligned timeseries root.sg.d1(s1 int32, s2 int32, s3 int32)");
+      statement.addBatch("insert into root.sg.d1(time,s1) values(3,1)");
+      statement.addBatch("insert into root.sg.d1(time,s1) values(1,1)");
+      statement.addBatch("insert into root.sg.d1(time,s1) values(2,1)");
+      statement.addBatch("insert into root.sg.d1(time,s2) values(2,2)");
+      statement.addBatch("insert into root.sg.d1(time,s2) values(1,2)");
+      statement.addBatch("insert into root.sg.d1(time,s2) values(3,2)");
+      statement.addBatch("insert into root.sg.d1(time,s3) values(1,3)");
+      statement.addBatch("insert into root.sg.d1(time,s3) values(3,3)");
+      statement.executeBatch();
+
+      try (ResultSet resultSet =
+          statement.executeQuery("select count(s1), count(s2), count(s3) from root.sg.d1")) {
+
+        assertTrue(resultSet.next());
+        assertEquals(3, resultSet.getInt(1));
+        assertEquals(3, resultSet.getInt(2));
+        assertEquals(2, resultSet.getInt(3));
+
+        assertFalse(resultSet.next());
+      }
+
+      statement.execute("flush");
+      try (ResultSet resultSet =
+          statement.executeQuery("select count(s1), count(s2), count(s3) from root.sg.d1")) {
+
+        assertTrue(resultSet.next());
+        assertEquals(3, resultSet.getInt(1));
+        assertEquals(3, resultSet.getInt(2));
+        assertEquals(2, resultSet.getInt(3));
+
+        assertFalse(resultSet.next());
+      }
     }
   }
 }
