+++ integration-test/src/test/java/org/apache/iotdb/db/it/aligned/IoTDBInsertAlignedValuesIT.java
@@ -48,7 +48,6 @@ public class IoTDBInsertAlignedValuesIT {
   @Before
   public void setUp() throws Exception {
     EnvFactory.getEnv().getConfig().getCommonConfig().setAutoCreateSchemaEnabled(true);
-    EnvFactory.getEnv().getConfig().getCommonConfig().setMaxNumberOfPointsInPage(2);
     EnvFactory.getEnv().initClusterEnvironment();
   }
 
@@ -243,46 +242,6 @@ public class IoTDBInsertAlignedValuesIT {
         assertEquals(1, resultSet.getLong(1));
         assertEquals(2.0F, resultSet.getObject(2));
         assertEquals(2.0F, resultSet.getObject(3));
-
-        assertFalse(resultSet.next());
-      }
-    }
-  }
-
-  @Test
-  public void testInsertComplexAlignedValues() throws SQLException {
-    try (Connection connection = EnvFactory.getEnv().getConnection();
-        Statement statement = connection.createStatement()) {
-      statement.addBatch("create aligned timeseries root.sg.d1(s1 int32, s2 int32, s3 int32)");
-      statement.addBatch("insert into root.sg.d1(time,s1) values(3,1)");
-      statement.addBatch("insert into root.sg.d1(time,s1) values(1,1)");
-      statement.addBatch("insert into root.sg.d1(time,s1) values(2,1)");
-      statement.addBatch("insert into root.sg.d1(time,s2) values(2,2)");
-      statement.addBatch("insert into root.sg.d1(time,s2) values(1,2)");
-      statement.addBatch("insert into root.sg.d1(time,s2) values(3,2)");
-      statement.addBatch("insert into root.sg.d1(time,s3) values(1,3)");
-      statement.addBatch("insert into root.sg.d1(time,s3) values(3,3)");
-      statement.executeBatch();
-
-      try (ResultSet resultSet =
-          statement.executeQuery("select count(s1), count(s2), count(s3) from root.sg.d1")) {
-
-        assertTrue(resultSet.next());
-        assertEquals(3, resultSet.getInt(1));
-        assertEquals(3, resultSet.getInt(2));
-        assertEquals(2, resultSet.getInt(3));
-
-        assertFalse(resultSet.next());
-      }
-
-      statement.execute("flush");
-      try (ResultSet resultSet =
-          statement.executeQuery("select count(s1), count(s2), count(s3) from root.sg.d1")) {
-
-        assertTrue(resultSet.next());
-        assertEquals(3, resultSet.getInt(1));
-        assertEquals(3, resultSet.getInt(2));
-        assertEquals(2, resultSet.getInt(3));
 
         assertFalse(resultSet.next());
       }
